requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
groq==0.9.0
langchain==0.3.0
langchain-groq==0.2.0
langchain-community==0.3.0
chromadb==0.5.5
python-dotenv==1.0.1
pydantic==2.9.2
httpx==0.27.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
python-multipart==0.0.9
sentence-transformers
pypdf
